PVW-1434: History detail fix
* Clean up TimelineAttributeRepository

As we are moving closer to the core certain (stale) mock
apis and usecases become unused. This commit cleans up the
timeline_attribute_repository and removes all the (now unused)
usecases that relied on these methods.

* Remove cardId from HistoryDetailScreenArgument

Remove the id since it's no longer used to perform
any filtering.

* Pass TimelineAttribute model on navigation

Since the core does not provide stable IDs for
WalletEvents (and thus TimelineAttributes) we now
pass on the complete TimelineAttribute model on
navigation to the detail screen.