Remove empty placeholder card from mock log
* Remove empty placeholder card

In the missing attribute flow a placeholder RequestedCard was created
to satisfy the api. However this caused the wallet_card_front mapper
to fail when any missing attribute wallet event had to be rendered
in the UI. This commit simply logs an empty list, which is (at least
for now) sufficient since the requested attributes are not shown in
the 'sharing stopped' ui state.