PVW-1429: Disclosure flow fixes
* Verify back button behaviour

* Verify disclosure happy path and reporting flow

* Update DisclosureBloc constructor

Instead of initializing the disclosure as soon as
the bloc is created by having the uri as a parameter
on instantiation, prefer passing it in as an event.
This allows us to test the bloc more properly (e.g.
removing a flaky Future.delayed) and in general is
better practice.

* Verify stopDisclosure from missing attributes state

* Add approve organization tests

* Clean up blocTest syntax

Use setUp to initialize the usecases

* Add basic DisclosureBloc Tests

* Add start/cancel disclosure usecase mocks

* Cancel disclosure on fatal pin error

When the user is confirming a disclosure session with a
pin, but fails to provide the correct pin (triggering a timeout)
the ongoing disclosure session is now cancelled when the bloc is
closed, to make sure the user can re-initiate a disclosure flow without
the app crashing.

* Handle DisclosureReportPressed

The disclosure bloc did not handle the DisclosureReportPressed
event yet, causing the app to crash when the user selected an
option from this screen. This if fixed by handling it explicitly
in the bloc.