Implement incrementing mdoc_copy disclosure_count
This adds a method to `Strorage` to increment the `storage_count` on
multiple `mdoc_copy` entries based on their `id`s. Included are tests
for this and the query behaviour implemented in the previous commit.

This also includes a differentiation in the two identifiers returned
for unique mdocs by `Storage`, as previously they were conflated.