PVW-1844: Re-enable card history
* Apply textStyle to timeline_section_header

* Fix TimelineSectionHeader height

It had some weird top padding due to an invalid
height calculation, fixed by calculating it
with a TextPainter.

* Remove unused OrganizationDataSource

* Fix CardHistoryBloc issue

The CardHistoryBloc wasn't able to fetch it's data
since it provided a docType to an api expecting the
cardId. Relevant apis that were still relying on the
cardId have been updated to rely on the docType now,
this fixed the issue when loading the history per card.

* Re-enable history per card

Tim requested this feature to be activated again,
as such re-enabling it. Since it was the only feature
flag we used, opted to remove the class completely.

* Fix CardHistoryScreen padding issue

On the error screen the retry button was stuck to the
bottom of the screen due to missing padding.