PVW-1895: Optional requested cards
* Fix (mock) history sorting

Since sorting now happens inside the wallet_core
we no longer explicitly sort in the wallet_app. This
caused the sorting to break for mock builds as it relied
on things being sorted on the wallet_app side. This commit
fixes this by making sure the correct sorting is provided
from inside the mock api.

* Make WalletEvent.requested_cards optional