PVW-1856: (Sliver)WalletAppBar Widget
* Update WalletAppBar elevation effect

Update the WalletAppBar to include the expected elevation
effect when the content is scrolled. Also updated the design
screen sample page to be able to test this behaviour.

* Use context extension

* Add leading field to WalletAppBar

* Introduce WalletAppBar

When the collapsing effect is not needed, and we can get away
without a CustomScrollView we can rely on the WalletAppBar
to draw the appbar at the top of the screen. It still supports
the progress indicator and styles the title etc. as expected.

* Introduce StepperIndicator

Introduce a widget that renders the stepper with the correct
side padding pre-applied. Also add the SliverWalletAppBar
to the design system screen so it can be verified more
easily.

* Refactor WalletAppBar to SliverWalletAppBar

Be consistent in prefixing all sliver widgets
with the word sliver.

* Update themes

Update the themes to reflect the
newly desired app bar styles.

* Introduce WalletBackButton

Drop in replacement for the BackButton but now with
the desired icon.

* Introduce WalletAppBar

Introduce the WalletAppBar, to be used as the first
entry in a CustomScrollView to create the behaviour
where the title, which is initially shown in-page,
scrolls into the app bar as the content is scrolled.