PVW-1739: Update Organization Pages
* Use context extensions

* Flutter analyze fix

* Lokalise interactions text on OrgDetailsScreen

* Fix OrganizationDetailScreen Test

* Update localizations

* Restructure info section

Extract into methods and make sure methods are
ordered according to UI order.

* Update OrganizationDetailScreen info section

* Display interaction row

* Update OrganizationDetailScreen and AppBar

* Add help button and remove title

* Update OrganizationApprovePage Design

* Introduce LinkTileButton

The LinkTileButton is a full width button similar
to the LinkButton, but providing a larger tap target
and visual feedback.
Also moved two other button widgets into the button
subdirectory.

* Update OrganizationApprovePage Logo

Tweak the borderRadius

* Remove commented out leading button

* Use context extension

* Remove potential back button from PinScreen

* Fix WalletPersonalizeScreenTest

* Make CardHistoryScreenTest use l10n

* Fix AboutScreenTest

* Fix CardHistoryScreenTest

Make it so the appbarTitleFinder expects two widgets
since the title is rendered in both the collapsed and
expanded state.

* Fix HomeScreenTest

* Remove unused strings

* Update WalletAppBar in PinTimeoutScreen

* Update WalletAppBar in PinBlockedScreen

* Update WalletAppBar in HistoryDetailScreen

* Update WalletAppBar in DataIncorrectScreen

* Update WalletAppBar in WalletPersonalizeSetupFailedScreen

* Update WalletAppBar in ThemeScreen

* Update WalletAppBar in ErrorScreen

Note: the ErrorScreen should eventually be
migrated to use the SliverWalletAppBar, but
due to the separation of the ErrorPage (excluding
the scaffold and app bar) and the ErrorScreen (which
includes the appbar) and the usage of both (e.g. the
ErrorPage is used inside the disclosure flow) this
change is not trivial and is therefor kept out of
this MR.

* Update WalletAppBar in DigidHelpScreen

* Update WalletAppBar in QrScreen

* Update WalletAppBar in WalletPersonalizeDataIncorrectScreen

* Update WalletAppBar in CardOverviewScreen

* Fix mock wallet.unlock() issue

Noticed that the unlock didn't work as expected
on mock builds due to an invalid pinMatches check.

* Update WalletAppBar in PinScreen

* Update _buildCloseButton

Clean up the buildClose button method for the
issuance and sign flows.

* Simplify DisclosureScreen changes

Use the same WalletAppBar implementation as the
issuance and sign screens, meaning we watch() the
progress to update the app bar.

* Update WalletAppBar in IssuanceScreen

* Update WalletAppBar in SignScreen

* Update WalletAppBar in CardDataIncorrectScreen

* Update WalletAppBar in CardDataScreen

* Update WalletAppBar in CardDetailScreen

* Update WalletAppBar to use Widget

Instead of providing a String title the title
parameter is now a Widget, so it can be provided
more dynamically. The same change is NOT made for
the SliverWalletAppBar because it relies on the
String value to perform certain measurements.

* Update theme for Ripple effect

The Elevated button had no 'tap' feedback because
the ripple color was identical to the button color,
fixed by changing the overlayColor of the elevated
button in the theme.

* Update WalletAppBar in WalletPersonalizeNoDigidScreen

* Update WalletAppBar in ForgotPinScreen

* Remove unused key: verifierPolicyScreenTitle

* Update WalletAppBar in PolicyScreen

* Update WalletAppBar in CheckAttributesScreen

* Update WalletAppBar in ReportIssueScreen

* Update WalletAppBar in OrganizationDetailScreen

* Show WalletAppBar in AboutScreen

* Show WalletAppBar in ChangeLanguageScreen

* Show WalletAppBar in SettingsScreen

* Show WalletAppBar in MenuScreen

* Update WalletAppBars

Take canPop into account to render the correct
back button when no leading widget is directly
provided.

* Add WalletAppBar to PlaceholderScreen

* Add WalletAppBar to DisclosureScreen

* Add WalletAppBar to CardHistoryScreen

* Add WalletAppBar to HistoryOverview

Migrated back to flutter_sticky_header for the
sticky headers because using the SliverMainAxisGroup
conflicted with the SliverWalletAppBar in some way,
causing the sticky headers to scroll behind the app
bar.

* Add WalletAppBar to leftover personalize pages

Some pages of the personalize flow were still missing the WalletAppBar
widget. This commits adds the WalletAppBar to those, completing the whole
personalization flow.

* Add WalletAppBar to WalletPersonalizeConfirmPinPage

* Add WalletAppBar to GenericLoadingPage

* Add leading field to WalletAppBar

* Introduce WalletAppBar

When the collapsing effect is not needed, and we can get away
without a CustomScrollView we can rely on the WalletAppBar
to draw the appbar at the top of the screen. It still supports
the progress indicator and styles the title etc. as expected.

* Introduce StepperIndicator

Introduce a widget that renders the stepper with the correct
side padding pre-applied. Also add the SliverWalletAppBar
to the design system screen so it can be verified more
easily.

* Refactor WalletAppBar to SliverWalletAppBar

Be consistent in prefixing all sliver widgets
with the word sliver.

* Update Appbar for WalletPersonalizeCheckDataOfferingPage

* Update Appbar for WalletPersonalizeIntroPage

* Update AppBar in setup_security_screen

* Update IntroductionConditionsScreen appbar

* Update IntroductionPrivacyScreen AppBar

* Update themes

Update the themes to reflect the
newly desired app bar styles.

* Introduce WalletBackButton

Drop in replacement for the BackButton but now with
the desired icon.

* Introduce WalletAppBar

Introduce the WalletAppBar, to be used as the first
entry in a CustomScrollView to create the behaviour
where the title, which is initially shown in-page,
scrolls into the app bar as the content is scrolled.