Configure tests_integration tests
Make the `db_test` feature explicitly required on all integration tests
contained in the `test_integration` crate.

Require any other crate that contains integration tests along with the
features required for this in the `test_integration` crate as a dev
dependency, except for `db_test`. This will cause the features to be
enabled through feature unification in the main workspace.