PVW-1937: Pass on and render organization privacy policy url
* Remove unused translation

Remove 'organizationDetailScreenPrivacyValue'

* Propagate privacyPolicyUrl from core

The privacyPolicyUrl was already available in the core,
this commit makes sure it propagates to the flutter
side as well.

* Introduce launch_util.dart

Since we want to avoid crashing when opening a
Uri or a Url string we always try/catch the launch
method. This commit replaces all those try/catch
blocks with a `launchUrlStringCatching` method in
the launch_util.dart file to do this consistently
throughout the whole app.

* Add privacyPolicyUrl to organization model