Add Organization to Card model
Update the Card model to contain the Organization
model which is now always provided by the core. Also
use this field to render the correct name on the
CheckAttributesScreen. It required some new generated
json code to make sure the Card model can still be serialized,
which might be used to pass it around between screens.