PVW-1749: ActivitySummary widget
* Improve ActivitySummary Test

* Add ActivitySummary tests

* Clean up ActivitySummary

Clean up the activity summary, splitting up the
line generation to make things a bit more readable.

Also added the new shared with case where the
organizations are grouped (e.g. '5 other parties')
when the user had interactions with >3 different
parties.

* Localise ActivitySummary

Add localizations for the ActivitySummary widget.

* Update ActivitySummary to generate title

Have the title of the ActivitySummary widget be
dynamically generated based on the provided
attributes.

* Introduce ActivitySummary widget

Introduce a widget that will show a summary
of the users history events.