PVW-1748: Dashboard content redesign
* Implement MR Feedback

- Align field ordering in TextWithLink
- Fix stale CardOverView namings

* Fix home_screen_test

Use l10n to resolve the expected widgets based
on their copy.

* Clean up DashboardScreen

* Add and use QR Button SVG

* Move dashboard classes to feature folder

* Rename CardOverview -> Dashboard

* Add TextWithLink to design system page

* Add TextButton below QR logo

* Introduce TextWithLink

Introduce a TextWithLink button to help render
texts with integrated links. Use this for the
dashboard footer and the organization approve
page for consistency throughout the project.

Also added a placholder (mostly generated) QR
code to the dashboard, to be replaced with the
proper logo.

* Restructure CardOverViewScreen

Basic restructure of the CardOverViewScreen to
render the skeleton of the new Dashboard.