PVW-1753: QR Scanner Redesign
* Implement MR Feedback

Refactor QrScanXx -> QrXx

* Menu button accessibility tweak

* Remove HomeTab.qr and fix tests

* Fix QR related tests

* Add QrScreen Localisations

* Remove unused QR classes

Removed:
- Flashlight cubit
- QrScanTab
- Flashlight proxy
- QrScannerFrame

Moved:
- QrScanBloc

* Fix arbs

* Lokalise Invalid QR Dialog

* Disable theme override

Use the device/app theme for the QrScreen instead
of forcing it to be in dark mode.

* Fix visibility of torch button

Use the hasTorch listenable to make sure the app
correctly rebuilds when the torch state becomes
available.

* Show Dialog on Invalid QRs

Show a dialog which informs the user when an invalid
qr code has been scanned.

* Update Loading & Success State

* Introduce QrNoPermission Widget

Implement a dedicated widget to handle the
no permission state of the Qr Scanner.

* Update QR Scanner Design

Update the design of the QrScanner widget and
use it to implement the QrScanScanning state.

* Force QrScreen to dark mode

* Make QrScreen fullscreen