PVW-1968: Show Activity on Dashboard
* Fix tests

The dashboard test needed some configuration to make sure
it didn't block on the VisibilityDetector change detection
timer. This commit configures it so it runs as expected
for all tests that might rely on the VisibilityDetector.

* Flip refresh logic

Instead of checking if the full widget is visible, refresh
as soon as some of it is visible, in case the visibleFraction
will be >0 but <1 on some (weird) devices.

* Refresh Dashboard when it becomes visible

To make sure the history is always in sync, we detect
when the Dashboard becomes visible with a Visibility
Detector and fire a refresh event accordingly.

* Remove required history from DashboardLoadSuccess

Make it more explicit when the history is empty vs unloaded, which
is a state that will happen when navigating to the dashboard with the
card animation, as in that case the preloaded cards have to be provided
to be able to perform the hero animation. The ui handles it the same
by providing an empty array to the ActivitySummary when the history is
null but this allows us to alternatively show a loading state if this
is desired at some point.

* Remove automatic dashboard loading on creation

It is considered bad practice to fire events from inside the
initialization method of a BLoC, as such moving this event to
where the bloc is initialized.

(bad practice because it can result in unexpected behaviour while
testing)

* Transform DashboardLoadTriggered events

Add the bloc_concurrency package and use the restartable()
transformer while observing the DashboardLoadTriggered events.
This makes sure that when we trigger the 'DashboardLoadTriggered'
multiple times, we only actually observe the walletCardsUseCase
stream once (i.e. any previous subscriptions will be cancelled).

Without this transformer we would potentially be observing the
ObserveWalletCardsUseCase multiple times.

* Update DashboardScreenTest

Moved from CardOverviewScreen Tests.

* Propagate TimelineAttributes to dashboard

* Add onTap to ActivitySummary