PVW-2123: Update Flutter dependencies
* Downgrade ffigen to 9.0.1

Newer version is incompatible with code generation
of flutter_rust_bridge.

* Run formatter and build_runner

* Apply suggested lint fixes (dart fix)

* Clean up CardHolograph

* Remove foil dependency

Foil widgets started causing the app to crash with all the other
dependency updates. To fix this crash I opted to roll my own implementation
of what foil achieves, which is painting a gradient on top of another
widget and slightly moving it based on the gyroscope input. This is
done using the new GyroGradientOverlay widget.

* Update Podfile.lock for new dependencies

* Update wallet_core dependencies

* Update wallet_app dependencies