PVW-1805: Update app introduction
* Center logo_rijksoverheid_label

Wrap the image in a Center widget so that the
screen reader highlight only draws a border around
the logo and not the complete width of the screen.

* Merge introduction page title and description

Merge the semantics so the title and description
are read out together.

* De-duplicate introduction buildPage logic

* Lokalise introduction_screen copy

* Update introduction placeholder images

* Add scrollbars to introduction

* Update placeholder image scaling

Use a SliverFillRemaining widget to make sure the
placeholder image scales to fill the viewport on
larger screen devices.

* Fix Title animation

Fading in at a more appropriate offset

* Remove IntroductionExpectationsScreen

Remove the expectations screen and add the illustration
placeholder.

* Pass on Keys

Make the ConfirmButtons widget support passing down keys
so that E2E tests can resolve to the correct buttons.

Note that I removed the 'introductionNextPageCtaText' to avoid
bloating the ConfirmButtons widget.

* Respect SafeAreas

* Generalize ConfirmButtons widget

Make the ConfirmButtons a bit more configure-able by allowing
the caller to set the theme of the buttons and generalizing the
naming to primary/secondary actions so it's also usable in non
deny/confirm scenarios such as the onboarding.

* Move WalletAppBar

Put it inside the scaffold vs inside the page.

* Rewrite introduction_screen

- Setup initial redesign
- Add semantic labels to stepper