Fix accept pid with invalid pin flow
When trying to accept the offered pid with the
incorrect pin multiple times the app went into
an invalid state because the 'reset' logic is now
implemented by the wallet_core, causing the app to
(invalidly) show the pin overlay. This is now fixed
by:
- Making it so that the WalletPersonalizeSetupFailedScreen
is no longer wrapped in a SecuredPageRoute, as the wallet is
now already reset before this screen is shown, so no need to
secure it with the pin overlay
- Removing the code that triggers the 'wallet reset' when the
dismiss button is pressed on the WalletPersonalizeSetupFailedScreen.
Doing so caused the app to crash as requesting a reset when the wallet
has already been reset triggers an invalid state exception.