Replace ConstructibleWithIdentifier trait
This replaces both the `ConstructibleWithIdentifier` and
`DeletableWithIdentifier` traits with a combined `StoredByIdentifier`
trait.

The previous implementation could lead to undefined behaviour if
multiple references to the same key exist and one of them deleted this
key. The new trait includes a contract where only one owner can exist
for a given key with an identifier. This behaviour is implemented in
both the hardware and software key structs.