PVW-2164: Expand wallet_app unit tests
* Update generated mocks

* Verify FlutterAppConfiguration.hashCode behaviour

* Update MissingAttribute test

Make sure the constructor is also tested.

* Add UiAttribute test

Also fix a bug where the 'key' was used in the equals
method, while the key should never be looked up in a
UiAttribute.

* Add MissingAttribute test

* Add LocalizedStringConverter test

* Add AttributeValueConverter test

* Add AppImageDataConverter test

* Verify NavigationService navigation events

* Verify contains pid behaviour

* Verify lockWallet exit behaviour

* Verify Legacy EdiQrCode behaviour

* Verify Legacy EdiQrCode behaviour

* Update CorePidRepository Tests

Make sure all methods are properly propagated to the
core library.

* Add basic AppLifecycleObserver test

Make sure the widget can be instantiated and renders
the provided child correctly.