PVW-2226 & PVW-2227: AndroidManifest updates
* Disable automatic backups in platform_support

Sonarqube raised a warning about the automatic backup
behaviour on android. Since we do not want to use this
functionality we explicitly disable it here.

* Disallow cleartext traffic

Sonarqube raised a warning that cleartext traffic was
(implicitly) enabled for older android versions. Now
explicitly disabling it to remove the warning as we are
not using cleartext (http) traffic.