PVW-2237: Propagate pin related (network)errors
* Rename NetworkErrorState filename

* Verify consumable behaviour

* Remove references to checking session

Leftover from when these changes were merged in a
single MR. Removing them (to be re-added later) for
now.

* Propagate pin errors during disclosure

* Add OnPinErrorCallback to PinPage

Add an explicit callback parameter to the PinPage so
that consumers of the PinPage can easily provide custom
handling for [ErrorState]s.

* Remove activeSession checks from wallet_core

Break up the PVW-2172 MR so that the rust changes can
be discussed and reviewed separately.

* Fix DisclosureBloc tests

Due to an extra call to cancel any ongoing disclosure
session, a slight delay is needed to avoid an unexpected
race condition in tests.

* Update WalletPersonalizeBloc

Update the WalletPersonalizeBloc so that it doesn't
deduce it's initial state from the provided pid
issuance argument, as this turns out to be a slight
anti pattern. Instead pass in the correct pid issuance
url right after the bloc creation (if an issuance url
is available).

* Consume pid issuance url

During pid issuance, if the wallet locks while
the user is looking at the offered pid data, the app
would crash when unlocking because the pid issuance auth
url is opened again (through continuePidIssuance()) because
the bloc is recreated with that as an argument. We now wrap
the url in a consumable so that the url is only used once.

* Update cancel pid_issuance / disclosure usecases

Update the cancel pid_issuance- / disclosure usecases to check for
active sessions before actually cancelling the flows. This avoids
state errors for incorrectly cancelling them. Ideally we should
still only cancel flows when appropriate but in certain cases (e.g.
opening 2 disclosure flows) we don't have a proper callback, so we
eagerly cancel the flow instead.

* Add session state APIs

Add apis to check the session state for pid
issuance and disclosure to the wallet_core api.

* Implement (Network)ErrorState consistently

Implement (Network)ErrorState consistently for
disclosure, pin and personalize. Also update
tests to verify new behaviour.

* Update ndkVersion in build.gradle

Update to ndkVersion "26.2.11394342"

* Propagate errors to WalletPersonalizeScreen

This commit makes it so that (network)errors during pin
entry (when accepting pid) are bubbled up to the
WalletPersonalizeScreen. To do so the (Network)ErrorState
classes are introduced which can be used to consistently
route errors through bloc states. To be consistent throughout
the app all error related states should be updated to implement
the corresponding (Network)ErrorState. This will be done in
a next commit.