PVW-2172: Add active session api to fix StateErrors
* Fix mutability of Wallet methods

* Add CancelDisclosureUseCase test

* Add CancelPidIssuanceUseCaseImpl test

* Update wallet_core_mock

* Fix PVW-2171

When the app locks while a pid issuance session is active,
the pid issuance flow will be restarted upon unlocking. This
means the state of the app and core are out of sync, to fix
this we eagerly cancel pid issuance when the user initiates
the digid login.

Considerations: Ideally we clean up the session in the close()
callback of the bloc, but cancelling requires the core to be
in an unlocked state, which it might no longer be when the
close callback is triggered. As such we eagerly call the cancel
api.

* Update WalletPersonalizeBloc for cancellation

Make sure the correct errors are propagated after
cancellation and that cancellation occurs on errors.

* Remove rejectOfferedPid from wallet_app

* Remove disclosure session from Wallet earlier

After successful disclosure, remove the session from `Wallet` before
attempting to story an event about in the history.

* Keep Wallet internal state on issuance state error

* Combine cancel and rejects method for PID issuance

* Wallet method tracing for issuance and disclosure

* Minor renames and fixes in wallet crate

* Remove unused network_error.dart

- Fix typo in translations
- Remove unused file
- Remove cancelPidIssuance (at least for now)

* Fix DisclosureBloc tests

Due to an extra call to cancel any ongoing disclosure
session, a slight delay is needed to avoid an unexpected
race condition in tests.

* Update WalletPersonalizeBloc

Update the WalletPersonalizeBloc so that it doesn't
deduce it's initial state from the provided pid
issuance argument, as this turns out to be a slight
anti pattern. Instead pass in the correct pid issuance
url right after the bloc creation (if an issuance url
is available).

* Consume pid issuance url

During pid issuance, if the wallet locks while
the user is looking at the offered pid data, the app
would crash when unlocking because the pid issuance auth
url is opened again (through continuePidIssuance()) because
the bloc is recreated with that as an argument. We now wrap
the url in a consumable so that the url is only used once.

* Update cancel pid_issuance / disclosure usecases

Update the cancel pid_issuance- / disclosure usecases to check for
active sessions before actually cancelling the flows. This avoids
state errors for incorrectly cancelling them. Ideally we should
still only cancel flows when appropriate but in certain cases (e.g.
opening 2 disclosure flows) we don't have a proper callback, so we
eagerly cancel the flow instead.

* Add session state APIs

Add apis to check the session state for pid
issuance and disclosure to the wallet_core api.

* Implement (Network)ErrorState consistently

Implement (Network)ErrorState consistently for
disclosure, pin and personalize. Also update
tests to verify new behaviour.

* Update ndkVersion in build.gradle

Update to ndkVersion "26.2.11394342"

* Propagate errors to WalletPersonalizeScreen

This commit makes it so that (network)errors during pin
entry (when accepting pid) are bubbled up to the
WalletPersonalizeScreen. To do so the (Network)ErrorState
classes are introduced which can be used to consistently
route errors through bloc states. To be consistent throughout
the app all error related states should be updated to implement
the corresponding (Network)ErrorState. This will be done in
a next commit.