Debounce lockedStream in PinOverlay
Instead of the elaborate logic to build only when
the app is in the registered state, which caused
multiple childs to be created this commit prefers
debouncing the locked state to allow any child
widget to finish processing events that might come
in the moment the app is locked externally (i.e. by
the wallet_core).