Replace debounce with registration check
Instead of debouncing to make sure any [child] widget is
kept on screen long enough to process any ongoing events
after the app is externally (by the core) locked. We now
combine the lock stream with a registration check, if the
wallet gets locked while its not yet registered we avoid
replacing the child with the pin screen, allowing the child
widget to continue processing any events as expected.