PVW-2349: Clean up `wallet_app` //FIXMEs
* Update errorScreen docs

* Uncomment disclosure cancellation on error

While the app always displays an error after _onConfirmPinFailed,
thus being forced to restart any disclosure session (which will
cancel any previous ongoing session anyway). I opted to uncomment
the cancellation here too. Mainly to clean up the session in the
edge case that there is still something active on the wallet_core
side, after which the user doesn't restart a session, which might
result in a long running session. Also, cancellation is preceded
by a check if a session is active, so calling it in this error
usecase will not have any negative side effect as it will simply
not be cancelled if it doesn't exist anymore.

* Update RedirectError.accessDenied related comment

* Remove 'confirmTransaction' api

Api was a leftover of the mock implementation and no
longer relevant.

* Clean up showExitSheet logic

* Update WalletCardItemTest docs

* Update PidAttributeMapper docs

Whether this is dynamic or not is ultimately up to
the UI/UX team and will then result in a ticket.

* Update CardFrontMapper docs

* Update WalletPersonalizeScreen docs

* Update OrganizationApprovePage docs

* Update IssuanceBloc docs

* Remove GetWalletCardUpdateRequestUseCase

Updating wallet cards is out of scope for the MVP, removing
the related usecase and thereby cleaning up a fixme and
explicit references to the mock.

* Remove dataContainsSingleViewProfilePhoto

This usecase was removed (now always false) and thus the property
and related business logic can be removed.

* Remove dataIsSignature from Policy

Instead we always show it inside the 'sign' flow, which
is the current usecase for the mock scenarios.

* Update identifyUri related //FIXME

Remove in favor of a simple note, as the code will
be revisited whenever IdentifyUriResult changes since
we switch on that enum right below it.