PVW-1746: Update pin screens
* Update TerminalPage bottom margin

* Extract kLeftoverAttemptsBeforeDynamicWarning

* Fix broken pin related tests

* Minor tweaks

- Update no DigiD icon and remove unused SafeAre
- Update some docs
- Update unlock translation

* Revive PinHeader

Update the PinHeader to match the new default usecase
of how text should be aligned in both landscape and portrait,
yet making it configurable to still support the centered variant
for the unlock screen.

* Update BottomButton docs

* Fix incorrect title size in SliverWalletAppBar

The _textHeightCache was not cleared properly on orientation
changes, which could cause the title to render incorrectly if
the amount of lines spanned by the title changed. Fixed by always
invalidating it in the didUpdateWidget callback (which is triggered
for orientation changes too).

* Move DisclosureScreen SafeAreas

Pull the disclosure screen safe areas down into the
pages where relevant, to align with the personalize
flow. Also slightly tweak how the SafeAre in the
TextIconButton behaves for a more consistent ui across
platforms (android doesn't have symmetric horizontal
viewInsets which caused some discrepancies).

* Always enable 'forgot PIN' button

* Add configurable keyboard color

Add the ability to configure the keyboard color
and set it to primary for the unlock screen.

* Fix WalletPersonalizeScreen SafeArea related issue

- Update the terminal pages to the new (non legacy) ones
- Move the SafeAreas down so it can be controlled individually
per page (for the personalize flow)
- Update TextIconButton to take horizontal safe area into account

* LeftAlign forgotPin button in landscape

* Update PinPage landscape layout

Update the PinPage so that it correctly renders
in when the phone is in landscape.

* Remove the SafeArea around the PinKeyboard

Remove the safe area in favour of letting the parent
handle it for more flexibility.

* Update BottomButton allowing it to grow

Instead of forcing a specific height, update the button
to be at least that height so that it can grow with large
font scaling.

* Add pinErrorDialogFinalAttemptBody l10

Add the final attempt copy and show it in the error
dialog when appropriate.

* Move 'forgot pin' button

* Use pinErrorDialog translations

* Add pinErrorDialog translations

* Render pin errors as dialog

* Remove title from unlock screen

* Update SetupSecurityCompletedPage

Migrate to the non-legacy terminal page.

* Update lokalisations and landscape layout

* Update SetupSecurity to show errors as dialog