Update TODOs
* Add story for TODO on distinct reader authentications

* Remove TODO on considering to report errors per mdoc

If the wallet encounters an error when computing the disclosure of some of the mdocs, then it cannot send all attributes that the RP asks for. In that case the RP is likely to abort, in which case the attributes that the wallet did manage to send were sent uselessly. Therefore in this case the wallet should be conservative, and just abort the entire session instead.

* Remove unused method DeviceSigned::new_signature()

* Test DeviceSigned::new_mac()

* Remove outdated TODO on country and nationality in mock PID data

* Add stories for some TODOs

* Remove TODO in verifier about error codes sent by the wallet

There is nothing to do here, the verifier already handles this in the only sensible way: return them to the caller.

* Remove TODO on key rollover/kid from JWT module

The functionality of this TODO mentions should be implemented by the caller of this module instead.

* Remove outdated TODO about mdoc referrer URL

We have already decided to do nothing with this.

* Remove outdated TODO on matching received against requested attributes

This is now done by ItemsRequests::match_against_response().

* Don't panic if X059 cert subject contains a non-string

* Remove TODO about salt in HKDF

At no point in ISO 18013-5 is a HMAC key derived without a salt.

* Remove outdated TODO about multiple issuer certs in mdoc COSE header