PVW-2503: Fix dismiss dialog bug
* Fix dismiss dialog bug

When the user entered an incorrect pin while the wallet was
locked, the error dialog did not correctly show up. This
happened because all dialogs get dismissed when the app gets
locked (this is desired behaviour). However if the app is already
locked this behaviour should not trigger again, which is what is
achieved with this commit.